requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-filter-stubs==0.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.1.0
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
paramiko==2.7.1
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pytz==2019.3
raven==6.10.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
uWSGI==2.0.18
